Create README.md
I have no idea if this is even right. I'm following the steps that I was given from the homepage after creating my account.

It said "Get Started with GIThub".

So I am hopefully doing the right things so far.

It mentioned starting with a README on step 2. 

Moving to Step 3. Add first Python File.